+++ edge-agent/src/Microsoft.Azure.Devices.Edge.Agent.Core/IRuntimeInfoProvider.cs
@@ -6,7 +6,6 @@ namespace Microsoft.Azure.Devices.Edge.Agent.Core
     using System.Threading;
     using System.Threading.Tasks;
     using Microsoft.Azure.Devices.Edge.Util;
-    using Newtonsoft.Json;
 
     /// <summary>
     /// This interface provides the module runtime information.
@@ -15,29 +14,10 @@ namespace Microsoft.Azure.Devices.Edge.Agent.Core
     /// </summary>
     public interface IRuntimeInfoProvider
     {
-        Task<IEnumerable<ModuleRuntimeInfo>> GetModules(CancellationToken ctsToken);
+        Task<IEnumerable<ModuleRuntimeInfo>> GetModules(CancellationToken cancellationToken);
 
         Task<Stream> GetModuleLogs(string module, bool follow, Option<int> tail, CancellationToken cancellationToken);
 
         Task<SystemInfo> GetSystemInfo();
-    }
-
-    public class SystemInfo
-    {
-        [JsonConstructor]
-        public SystemInfo(string operatingSystemType, string architecture, string version)
-        {
-            this.OperatingSystemType = operatingSystemType;
-            this.Architecture = architecture;
-            this.Version = version;
-        }
-
-        public string OperatingSystemType { get; }
-
-        public string Architecture { get; }
-
-        public string Version { get; }
-
-        static SystemInfo Empty { get; } = new SystemInfo(string.Empty, string.Empty, string.Empty);
     }
 }
